pyproject.toml
[project]
name = "biaas"
version = "0.1.0"
description = "This project is a BI as a Service that utilizes data from the Valencia City Hall API to generate reports and visualizations. The application processes the data in real-time and provides relevant insights on various aspects of the city."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.14.3",
    "dependency-injector==4.48.3",
    "faiss-cpu==1.13.2",
    "google-genai==1.64.0",
    "groq==1.0.0",
    "ollama==0.6.1",
    "plotly==6.5.2",
    "pydantic-settings==2.13.1",
    "sentence-transformers==5.2.3",
    "streamlit==1.54.0",
    "torch==2.10.0",
    "torchvision==0.25.0",
]

[dependency-groups]
dev = [
    "black==26.1.0",
    "mypy==1.19.1",
    "pandas-stubs==3.0.0.260204",
    "plotly-stubs==0.1.2",
    "ruff==0.15.2",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = "pydantic.mypy"

[[tool.mypy.overrides]]
module = ["faiss.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
